# backend/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "models.CreateTransactionRequest": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "purpose": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/models.TransactionType"
                    },
                    "wishlist_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "amount",
                    "category",
                    "date",
                    "type"
                ],
                "type": "object"
            },
...
            "models.Transaction": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/models.UserInfo"
                    },
                    "date": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "detail_request_id": {
                        "type": "string"
                    },
                    "detail_request_status": {
                        "type": "string"
                    },
                    "excluded_from_stats": {
                        "type": "boolean"
                    },
                    "hidden": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "string"
                    },
                    "last_modified_by": {
                        "$ref": "#/components/schemas/models.UserInfo"
                    },
                    "parent_id": {
                        "description": "Detail-request linkage. ParentID is set on child transactions; it\nreferences the lump-sum parent. DetailRequestID/Status are set on the\nparent transaction once a detail-request is opened over it.\nExcludedFromStats is denormalized so aggregation queries stay simple:\nwhile a request is open, children are excluded; once closed, the parent\nis excluded and the children become canonical.",
                        "type": "string"
                    },
                    "purpose": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/models.TransactionType"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer"
                    },
                    "wishlist_id": {
                        "description": "WishlistID links an expense transaction to a recurring wishlist item.\nEmpty when the transaction is not a recurring-payment fulfillment.\nForecast aggregation uses it to decide whether a recurring item is\n\"paid this period\" and should be excluded from the projected total.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "models.TransactionType": {
                "enum": [
                    "income",
                    "expense",
                    "initial_balance"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Income",
                    "Expense",
                    "InitialBalance"
                ]
            },
...
            "models.UpdateTransactionRequest": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/models.UserInfo"
                    },
                    "date": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "hidden": {
                        "type": "boolean"
                    },
                    "purpose": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "wishlist_id": {
                        "description": "Pointer so an empty string (\"\") is distinguishable from absent — empty\nstring unlinks the transaction from its wishlist item.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "models.UserInfo": {
                "properties": {
                    "avatar_url": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Заголовок вида `Bearer \u003cjwt\u003e` (получить через `/auth/login`).",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "msdnna",
            "url": "https://github.com/msdnna/budget"
        },
        "description": "REST API семейного бюджет-приложения (Go + MongoDB). JWT (Bearer) — защищённые ручки требуют заголовок `Authorization: Bearer \u003ctoken\u003e`, полученный через `POST /api/auth/login`.",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "title": "msdnna Budget API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/transactions": {
            "get": {
                "description": "Фильтры комбинируются по И. `categories` — список через запятую (приоритетнее `category`). `include_detailed=true` показывает родителей закрытых detail-requests.",
                "parameters": [
                    {
                        "description": "income|expense|initial_balance",
                        "in": "query",
                        "name": "type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Имя категории",
                        "in": "query",
                        "name": "category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "CSV категорий",
                        "in": "query",
                        "name": "categories",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "YYYY-MM-DD",
                        "in": "query",
                        "name": "from",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "YYYY-MM-DD",
                        "in": "query",
                        "name": "to",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Страница (1-based)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Размер страницы (1-100)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Показывать родителей закрытых detail-requests",
                        "in": "query",
                        "name": "include_detailed",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Список транзакций с пагинацией",
                "tags": [
                    "transactions"
                ]
            },
            "post": {
                "description": "Создаёт доход / расход / начальный баланс. `date` принимается в формате YYYY-MM-DD.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.CreateTransactionRequest",
                                        "summary": "body",
                                        "description": "Тело"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Тело",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.Transaction"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Создать транзакцию",
                "tags": [
                    "transactions"
                ]
            }
        },
        "/transactions/{id}": {
            "delete": {
                "description": "Запись помечается `deleted_at`; синк-клиенты увидят её при следующем pull.",
                "parameters": [
                    {
                        "description": "Transaction ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Удалить транзакцию (soft-delete)",
                "tags": [
                    "transactions"
                ]
            },
            "put": {
                "description": "Patch-семантика — пустые поля игнорируются, кроме `wishlist_id` (пустая строка отвязывает от wishlist-итема).",
                "parameters": [
                    {
                        "description": "Transaction ID (UUID)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.UpdateTransactionRequest",
                                        "summary": "body",
                                        "description": "Поля для обновления"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Поля для обновления",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.Transaction"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Обновить транзакцию",
                "tags": [
                    "transactions"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api"
        }
    ]
}